requirements/development.txt
-r base.txt
django-debug-toolbar>=4.3
django-browser-reload>=1.12
ruff>=0.4
pytest-django>=4.8
factory-boy>=3.3
